Inline MemCategorization into ExprUseVisitor
